graph summary
