+++ src/components/Graphs/GraphBody/graphBody.tsx
@@ -1,0 +1,13 @@
+import GraphSummary from "../GraphSummary/graphSummary";
+export default function GraphBody() {
+  return (
+    <div className="graph-body-container">
+      <span className="graph-body-wrapper">
+        <GraphSummary />
+      </span>
+      <span className="graph-body-wrapper">
+        <GraphSummary volume={true} />
+      </span>
+    </div>
+  );
+}
